requirements.txt
alembic==0.7.2
ecdsa==0.11
Fabric==1.10.1
Flask==0.10.1
Flask-Bootstrap==3.3.0.1
Flask-Menu==0.2.0
Flask-Migrate==1.3.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.0
MarkupSafe==0.23
paramiko==1.15.2
psycopg2==2.5.4
py==1.4.26
pycrypto==2.6.1
Pygments==2.0.1
PyHamcrest==1.8.1
pytest==2.6.4
six==1.8.0
SQLAlchemy==0.9.8
urwid==1.3.0
Werkzeug==0.9.6
WTForms==2.0.1
